+++ Reiseplaner/app/src/main/res/drawable/ic_add_floatingaction_btn.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#444444"
+    android:viewportHeight="24.0" android:viewportWidth="24.0"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M19,13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z"/>
+</vector>
